+++ include/Engine/Rendering/ModelJob.h
@@ -62,7 +62,7 @@ struct ModelJob : RenderJob
         if (world->HasComponent(Entity, "Animation") && Skeleton != nullptr) {
             auto animationComponent = world->GetComponent(Entity, "Animation");
             Animation = model->m_RawModel->m_Skeleton->GetAnimation(animationComponent["Name"]);
-            AnimationTime = (double)animationComponent["Time"];
+            AnimationTime = (float)((double)animationComponent["Time"]);
         }
     };
